+++ functions/api/[[route]].ts
@@ -1,5 +1,5 @@
 import { handle } from "hono/cloudflare-pages";
 
-import app from "..";
+import app from "../server";
 
 export const onRequest = handle(app);
